Restructured validate_config.R
This is now an outward-facing function, calling an internal function, in turn calling multiple separate testing functions.
The testing functions each raise "messages" for invalid conditions. These are (invisibly) collected by the internal function. If the config file is invalid, the external function raises an error and additionally prints the collected messages.